updated deprecated wx.Python, trimmed code.
