+++ README.md
@@ -17,7 +17,11 @@ jamarrange is built on python 2.7 More specifically 2.7.12. It should work fairl
 jamarrange utilises WXpython for it's GUI capabilities and functionalities. To develop using WXpython modules, simply download and setup WXpython for mac as instructed for your machine and the specific python version you are running on the [following link](https://www.wxpython.org/download.php).
 
 ##### Tinytag
-jamarrange relies on the Tinytag plugin to handle media file metadata. The installation details for setting up and installing TinyTag on python 2 can be found on the [following link](https://pypi.python.org/pypi/tinytag/).
+jamarrange relies on the Tinytag plugin to handle media file metadata. The installation details for setting up and installing TinyTag on python 2 can be found on the [following link](https://pypi.python.org/pypi/tinytag/) or by installing it through pip install, as illustrated below:
+
+```
+pip install tinytag
+```
 
 Tiny tag is a library that allows you to access file meta data (mainly audio files within jamarrange). It supports:
 - MP3
@@ -31,40 +35,46 @@ Tiny tag is a library that allows you to access file meta data (mainly audio fil
 
 To begin (once installed). Simply import your library and use the TinyTag.get() function to get all the metadata contained within a desired file. Use the file local directory as a string input as illustrated below.
 
-    from tinytag import TinyTag
-    tag = TinyTag.get('/some/music.mp3')
-    print('This track is by %s.' % tag.artist)
-    print('It is %f seconds long.' % tag.duration)
+```
+from tinytag import TinyTag
+tag = TinyTag.get('/some_folder/favourite_song.mp3')
+print('This track is by %s.' % tag.artist)
+print('It is %f seconds long.' % tag.duration)
+```
 
 Listed below is a list of possible attributes you can get with TinyTag:
 
-    tag.album         # album as string
-    tag.albumartist   # album artist as string
-    tag.artist        # artist name as string
-    tag.audio_offset  # number of bytes before audio data begins
-    tag.bitrate       # bitrate in kBits/s
-    tag.disc          # disc number
-    tag.disc_total    # the total number of discs
-    tag.duration      # duration of the song in seconds
-    tag.filesize      # file size in bytes
-    tag.genre         # genre as string
-    tag.samplerate    # samples per second
-    tag.title         # title of the song
-    tag.track         # track number as string
-    tag.track_total   # total number of tracks as string
-    tag.year          # year or data as string
+```
+tag.album         # album as string
+tag.albumartist   # album artist as string
+tag.artist        # artist name as string
+tag.audio_offset  # number of bytes before audio data begins
+tag.bitrate       # bitrate in kBits/s
+tag.disc          # disc number
+tag.disc_total    # the total number of discs
+tag.duration      # duration of the song in seconds
+tag.filesize      # file size in bytes
+tag.genre         # genre as string
+tag.samplerate    # samples per second
+tag.title         # title of the song
+tag.track         # track number as string
+tag.track_total   # total number of tracks as string
+tag.year          # year or data as string
+
+```
 
 To access file cover images from ID3 tags you can follow the example below:
 
-    tag = TinyTag.get('/some/music.mp3', image=True)
-    image_data = tag.get_image()
-
+```
+tag = TinyTag.get('/some/music.mp3', image=True)
+image_data = tag.get_image()
+```
 
 ## File arrangement
-Identify your music files and arrange them accordingly in a desired local storage folder. Using audio fingerprinting technology, the application accurately identifies artists, albums, and songs to accurately arrange your files.
+jamarrange accurately combs through all your file metadata and uses it to accurately arrange your music.
 
 ## Contribution
-Want to contribute? Simply clone the latest master repository (and version) and create a branch under your name.
+Want to contribute? Simply fork the latest master repository (stable release and version).
 
 ---
 # About
