--- filerepo/cleaner.py
@@ -1,2 +1,0 @@
-import shutil
-shutil.rmtree('C:\\Users\\User\\Music\\rAY\\Arranged files')
