--- mediaplayer/gridtest1.py
@@ -1,10 +1,0 @@
-from Tkinter import Tk, Label
-from ttk import Frame, Button, Style, Progressbar
-
-root = Tk()
-
-Label(root, text="File").grid(row=0)
-Label(root, text="Location").grid(row=1)
-
-
-root.mainloop()
