--- mediaplayer/model.py
@@ -1,4 +1,0 @@
-
-class Model:
-    def __init__(self):
-        pass
