--- mediaplayer/player.py
@@ -1,5 +1,0 @@
-import pyglet
-
-class Player():
-    def __init__(self):
-        pass
