--- tutorials/acoustid.md
@@ -1,3 +1,0 @@
-# Jam Arrange | Acoustid tutorial
-
-temporary API Key: AxIwHZrcAz
